Rev A.1 Functional release
Little to no commenting completed.
The code is functional, but I have several minimum improvements before using this for the assignment.
Mainly, overwriting the readout of what is currently playing. 

The sound files are added so a user can test the script without having to make edits and are free downloads from several sound effect sites.

Currently, any songs to be played must be stored in the same folder, and be of .mp3; .wav; or .ogg file types. ogg is untested. The file name must be manually entered into the list 'sound_files' verbatim as well.